refactor: extracted translation logic to separate module
